Clear old entry files
This is a potential solution to #3, obviously can use improvements, like support for [hash:10] (any custom hash length number), options which differ from [hash] like chunkhash etc.

I'm not entirely convinced this would be the way to go, let me know what you think? If you believe this could potentially fix things we can expand it, for me this is working rather nicely (I was debugging this.js to see if it clears old files).